Working PI-controller using node 1 ADC inputs.

+++ Node 2/include/input.h
@@ -4,8 +4,8 @@
 #include <stdint.h>
 
 typedef struct {
-  int8_t x;
-  int8_t y;
+  int x;
+  int y;
   bool btn;
 } joystick_xy_t;
